Move InitializationData to its own file
